removed old dropdown
