+++ src/components/Dropdown/Dropdown.scss
@@ -1,6 +1,6 @@
 @import '../../constants/styles';
 
-.old-dropdown {
+.Dropdown {
   display: flex;
   user-select: none;
   justify-content: space-between;
@@ -9,7 +9,6 @@
 
   border-radius: 4px;
   height: 28px;
-  width: 96px;
   border: 1px solid	var(--border);
   padding-left: 8px;
   padding-right: 4px;
@@ -40,7 +39,6 @@
     right: 0px;
 
     border-radius: 4px;
-    width: 94px;
     box-shadow: 0px 0px 3px 0px var(--document-box-shadow);
 
     background: var(--component-background);
@@ -53,6 +51,11 @@
       align-items: center;
       display: flex;
       border-radius: 0px;
+      padding-right: 32px;
+
+      &.active {
+        background-color: var(--popup-button-active);
+      }
 
       &:hover {
         background-color: var(--popup-button-hover);
@@ -70,7 +73,7 @@
     }
 
     &.hide {
-      display: none;
+      visibility: hidden;
     }
   }
 }
