--- src/components/Dropdown2/index.js
@@ -1,3 +1,0 @@
-import Dropdown from './Dropdown';
-
-export default Dropdown;
